Intial commit - Tree structure of a sign is mostly there
This is the base of my work for the revamped "Girouette" software. I
wanted to take back this piece of code and improve it, but it turned out
to be a nightmare (much of the actual processing was done in the GUI
functions, the structure of config / save files was very hard to
modify). The amount of work needed to have something cleaner would be
somewhat equivalent to writing the software again from scratch, so here
I go.

The goal here is to have a sign editor that's capable of (in order of
implementation):
1) its base feature: rendering signs "à la" Helen, Edigir, or others
(yet I don't have any such software that only bus operators have) from a
text, structure and font specification,
2) handling flip-discs, LEDs (monochrome and RGB),
3) animating LED signs (marquee, blink, enter left/right/top/bottom, or
any combination of these), and animating flip-disc display refreshes;
this comes with the ability to save to a GIF file.
4) handling sequences of animated signs (say, alternating route
destination and "Please stamp your ticket" messages).

The current status is: The tree structure defining a sign is mostly
there. There is an embryo of a render() function in every class making
up the tree, that's being moved to a visitor which will handle the
rendering alone.

The next milestone is to get a first (static) sign rendered: for that,
we need to add font support (perhaps first by taking the file format
used in Girouette) and implement the actual rendering process
(SignCellLeaf will be the hard part).