Image rendering progress, tabs become spaces
This commit introduces a sign renderer visitor class, plus an early
Bitmap class that can only render to a RGB32 image (which is accepted by
a QImage). A very early GUI is there to visualize the result; this
little incursion into the Qt world will help us throughout the
development of Siganim.

Some details to be remembered for later stages:

For the time being, cells are rendered as a black-and-white
checkerboard; we should state some clear specification of the color
palettes needed for all types of displays.

Also missing is placement information for the different displays (how
fine-grain should it be: just above/below/left-of/right-of? Absolute
coordinates (in which unit)?).