Change in structure - no more free-floating cells
This commit brings in a new sign structure: instead of having
free-floating cells that we can move around, we are going to have split
cells at a given position. Splits can either happen vertically or
horizontally.

This induced a deep change in how the sign tree is constructed: we no
longer need a SignTree class, instead we have a SignCell class that
restricts itself to those elements that have a height / width that can
be expressed in pixels. In particular, the Display class can have no
SignCell parent, but the tree structure is otherwise browsable. Browsing
up is read-only (const), only downwards browsing allows to modify the
objects.