More sound colors and images - color management undecided
There is no more link between SignImage and Bitmap. It's up to the sign
renderer to figure out how a sign dot should be converted to real-world
pixels: merely copying, applying a mask and a color, and so on.

The color system is far from perfect. The base idea of it is that:
  1) SignImages are independent of the type of display (whether it's
flip-disc, monochrome or RGB LEDs), so every dot must carry an on-off
(background-foreground) semantic along with an RGB value if that value
directly comes from what's being rendered. So, for flip-disc and
monochrome displays, whether the dots have an RGB value or not does not
matter, only their background-foreground status will.

  2) There are default RGB values provided globally for renderers to map
background and foreground semantic values to actual RGB values,

  3) For certain display types, the user should be able to input custom
RGB pixel values.

  4) Default values should be provided to all renderers, or all
renderers should have their own defaults (undecided, but one of these
has to happen).

There is one big issue for now, which is where should color semantic to
RGB conversion be performed. The current revision has it in
SignColor::getValue(), yet it seems to be more of the renderer's job to
have this.