Font import, renderer progress and a bit of GUI
This commit brings a bunch of essential features to Siganim.

1) Font support is added, with basic functionality (a font is a mere
container for some characters). One font parser adds support for font
files from Girouette, expect other formats to be supported later (so
more parsers to be added).

2) The renderer is actually rendering characters instead of a
checkerboard. Thus, we can have our very first signs on screen! (Very
exciting indeed to get there, because that's not so far from where
Girouette went). Yet, is it the renderer's (as visitor) role to render
individual characters? Wouldn't it be wise to split the renderer into
sub-modules (fonts, animations, ...) while letting it do display
division (cells) by itself?

3) Very little improvements have been done to the Qt GUI, just enough to
add a text zone which will be useful for testing. The GUI is not a
priority as of today (let's be clear on all the data structures first,
and then the GUI will have to reflect those choices).