Font family and style lookup, add GIF output support and const visitors
This commit brings in a GIF output support through a very simple library
(GIF-h). If that library proves insufficient, we may switch to GIFlib
later on.

The renderer is now a const visitor, which means it can't mutate the
nodes of a sign tree. There are now two visitor classes : a
ConstSignTreeVisitor, and a SignTreeVisitor in which the tree is
mutable.

We now also have font lookup by family and style, which will enable us
to have several variants of a font, just like modern operating systems
do.