One-way marquee animation support, fix cell type call
This commit introduces a sample animation (right-to-left marquee). It
was surprisingly easy to get this rendered, though a bit of adjustments
need to be done on making the marquee loop fluid.

The renderer visitor uses an extra structure called a sign tree, which
is used to place sign images before they are all merged at once. In
order to get animations running, we had a choice between modifying this
data structure to add shift nodes, or to do partial merges, doing the
shift while merging. We went the latter route, so we can duplicate the
animated image and make a nice loop.

This commit also removes the quick-and-dirty "Cell type string" static
function to delegate it to the class itself. That one is used whenever a
"cannot set parent" exception is encountered, so right now, only with
displays.