Support for text vertical/horizontal alignment, fix marquee break
This commit adds support for horizontal and vertical alignment. While
horizontal alignment is classical (left, center, right, justify),
vertical alignment allows the user to position the biggest character
(top, center, bottom) and the other characters relative to the biggest
one (top of the biggest, center of the biggest, bottom of the biggest).
This feature is usually not present in text editors, but is useful on
one-line displays where there are possibly big capitals alternating with
regular letters.

This commit also fixes the "break" in the marquee animation, and
introduces marqueeing both ways (left-to-right and right-to-left).