Blink animation, enhance image color semantics, lazy cropping
This commit introduces a blink animation. It is defined by the number of
on / off frames (there could be a total duration + duty cycle metric as
well, this has to be clearly defined).

Sign images now bear background color semantics. When they are composed
(split), that color becomes the default background, but that color is
supposed to be kept through animations.

Also, sign images can be wider / taller than the cell they are supposed
to fit into. The actual crop happens at compositing (or, if no
compositing is to happen, before the image is turned into a bitmap).
This allows to have marquee on text that overflows its box (which is one
of the goals of marqueeing text).