Add Compose node
This commit introduces a "compose" node into the sign cell structre.
This node is used to compose a foreground cell (for instance an animated
text) with a background cell (for instance a background color). By
adding an extra "fill" node into the tree, we will be able to get rid of
the text's background color property and transfer it to dedicated cells
(which will set the background color all the way up to a composite
node).