Add Fill cell, completely revamp color management, add defaults class
This commit adds a Fill cell, that can be used to set a fixed background
color to any other cell by composition. In order for this to work, we
have completely reviewed the color structure: a color now bears an
on/off semantic value, and an optional RGB value. When rendering, the
renderer will either hit a custom value and can use that one, or has to
provide a default color (an exception will be raised).

Default colors are provided as public members of a defaults structure.
This can stay until we find a better way of dealing with defaults.

The current renderer accepts custom RGB values for RGB LED signs only;
flip-disc and monochrome LEDs only use the binary on-off value of each
dot for rendering.