GUI progress, add observer pattern & sink
This commit brings up a tabbed GUI, with a renderer that's using the
Observer design pattern. Two observers are actually used : a renderer
sink (akin the GIF sink) watches a sign for changes, and a Qt widget
watches the renderer for changes. The renderer sink (ObservableSink) has
a permanent thread that pushes frames to the GUI on a periodic basis,
and if necessary (i.e. if the sign has changed), will re-render all
frames.

Thanks to these observers, the Qt GUI can now display animated signs and
be updated as they change.

In order for this to work, this needs that we be clear on the tree
hierarchy and especially be sure that we never reach out for a NULL
parent, and that every node has the right parent at the right time.
Quite a bunch of files have been changed in this regard.

Regarding the GUI itself, two widgets are in the GUI as "main tabs" : a
sign editor and an (empty) font studio. Font design ability is essential
for such a piece of software.