Cleaner parent-child relationship using exceptions
The setParent() method in each cell calls this->modified(), which is its
expected behavior. If the node is orphan, this now raises an exception
instead of using a possibly null pointer as reference to parent.

A similar process should be implemented when traversing the tree (throw
exceptions on getChildX() instead of having the receiver check for a
null pointer).