Add tree model in GUI, structure observer
This commit adds a tree widget in the sign editor GUI, representing the
sign tree. We use Qt's TreeWidgetItem as model, which is fairly
sufficient (we just want the tree widget to be aware of the tree nature
of the tree, it doesn't need to represent any other property). The
TreeWidgetItem is a structure observer (yet another kind of observer)
relative to the node it watches. When that node gets updated, it calls
the observer (Qt's tree) so that it reflects the update.

This commit also introduces the SignTree class once again. Here the
class is accompanied by its observer (the class itself is just worth
calling the observer back, giving it the opportunity to identify the
concrete class of the observed object).