New editor pages classes, separate observer from callback
This commit brings in per-class (empty) editors for sign tree nodes.
These will be used to edit the sign, so they are specific to one class
of nodes. This class is determined through a dispatcher callback, which
has been separated from the observer introduced in SignTree.

There are a lot of new files, yet few actually new content in this
commit...