Progress on some UI pages, more use of exceptions
This commit brings in a few new UI pages (split, text, empty cell) that
allow the user to see what they can do, or to set options if there are.

Being able to edit the structure live while it is on preview created a
new issue: we need to be able to cancel the rendering if the sign is
being modified while it is in progress. We'd need to put some mutex on
the tree and have the renderer work on a const deep copy, so that no
frame could ever get corrupted or, worse, lead to a crash (say, by a
child that disappears in the middle of it being rendered). During the
time this deep copy is created, the UI should be blocked so that no
further changes are possible.