More GUI pages, new layout, obs.sink & renderer progress, mem management
This commit comes with progress on the GUI, as almost everything is now
configurable, except the text's font (expected to be changed to a
string) and its vertical alignment (looking for the right way to expose
this to the user).

The GUI main layout has changed towards what will be the definitive
window layout (tree on the left, sign/cell edition on the right, sign
preview below).

The observable sink has been improved to not freeze the GUI for the
whole time the sign is being rendered. It will only be frozen for how
long it takes to make a deep copy of the sign on which the renderer's
thread can work hassle-free. Even though this comes with quite a bit of
bugs (it's easy to make the program crash by adjusting the display size
too many times in a row to something too big). Crashes are more likely
due to memory leaks than anything else, as very few efforts have been
done to address memory management until now. Trees are starting to have
a sound deletion policy (a sign tree node deletes all its children along
with it). The same kind of policy should also apply to the renderer's
SignTree and other structures, yet no action has been taken so far.