Switch to model/view for tree, completely mess up everything
What happened to me so that I chose to go down this road? This commit
brings in a Qt "model/view" pattern for the tree view of our SignTree
structure. Before this change, the tree was a QTreeWidget that just
happened to work, yet I was not happy with having to cast pointers to
its elements (that are not supposed to be subclassed, yet they were),
mainly in order to get context menus.

I tried to do things properly: I thus needed to use a model/view design
pattern, but not any kind of such. Qt forces on you how *they* think
such a model should be implemented: your model *has* to be representable
as a grid or a nest of grids and give out "index elements" that have no
pointers / references to their siblings or parents, yet have to be able
to tell who those folks are. The tree view will access your model as
such and point to your items with rows and columns.

Of course, we do not have such an ill-defined representation in our
tree; moreover, we do need extra items in the displayed tree that are
not present in the SignTree (empty nodes that can be selected to add new
nodes in their place, as not every node supports adding children). Thus
I introduced the IRPointer data structure that can accomodate a pointer
to either a SignTree, or one of "builder" classes that contain all
information to build a node later on. Plus, a bunch of maps and vectors
to build up an intermediate representation as required by Qt (a table
for parent "indexes", a table for children "indexes", a backwards map to
be able to get "indexes" from IRPointer*s...). No one imagines the
amount of crashes due to mismanagement of these data structures and the
time it took to debug that. And guess what, after 24 hours of work,
we're still not done! We can still observe nodes that do not select
themselves in the tree when clicked, or that select another node (which
means they have the same index or something like that).

If you have read the commit message up to this point, and you're willing
to do something to oust this f****d-up design pattern, the least I can
do is to get you a beer. If you spot a bug and give me its root, I owe
you two...