Progress on tree model, nodes can be added and removed, except display
This commit introduces a new proxy IR for sign trees, to replace the set
of lists we had in the last commit. This IR is a
super-double-linked-list where everyone has a pointer to its
predecessor, its successor, its parent and its first child. All of those
are updated by newly created or soon-to-be-removed nodes.

The difficulty that appeared with the table-based IR is that when a tree
node was added / deleted, other existing nodes would not get updated,
thereby returning possibly invalid indices. Qt documentation clearly
states that these indices are *transient* and should be discarded once
used; thus, we now build them on-the-fly, at the expense of browsing
this double-linked-list structure where we used to perform a simple
table lookup.

It looks like this new IR fits pretty well what Qt expects: nodes can be
added and removed dynamically. There remains an issue with top-level
nodes (try to remove a display and see what happens), because their
parent has no parent and so have to be looked up in some other way. I
think this will be solved by searching in the pointer vector (yes, we
keep a list of all the nodes we have) those nodes that happen to have no
parent. We would give them to Qt as top-level elements, instead of using
an arbitrary root node. Doing so will probably make our life easier when
multiple signs will be supported (hopefully, one day!).