Dirty fix for toplevel element deletion in GUI tree
The displayed tree in the GUI has a unique root node, which has a
pointer in the model (model->treeAsChain). Yet, both the SignTree
structure and the chain structure allow to have multiple orphan nodes;
thus, it is not wise to keep an unique parent. In order to replace this
unique node by a set of nodes (we have a vector for this purpose
already), we would need to be able to keep their relative order in case
a top-level node gets updated. The current procedure is that an updated
node's parent is deleted, then created again; vector lookup would mean
that should the first element be deleted, the node that will replace it
would be placed (and displayed) at the end of the tree.