Add font support in sign editor, change text node structure
This commit brings in font support in the text page of the font editor.
In order to achieve this, it also changes numerous constructors (all the
way from the main window to the text editor page) to pass on a font set
object. The user is then able to browse through fonts and pick that of
their choice through a combo box.

We also changed the text node structure in such a way that it contains a
text reference to its font, and not an actual pointer to its font. This
loose binding allows the creation and rendering of signs with fonts
absent from the current font set. This will also make it easier to
generate sign files, which format is still to be defined (JSON?).

A good idea would be to have a common format for signs, fonts and
anything else Siganim supports. Girouette used to have XML, which is
easy to parse, yet a bit hard to handcraft. JSON has enough expressivity
for us to specify our whole tree and font structures, and looks easy to
parse. We wouldn't even need to use a parser generator and a full-blown
DSL for that.

This commit finally removes the test file that's a burden to update and
no longer useful (there are too few features to be tested and no
good/easy way to check the renderer's result).