Font Studio progress, new Rasterizer renderer, new single frame sink
This commit brings in lots of improvements in the Font Studio. There is
now a nice list of characters which shows a preview of each of them
along with their Unicode name. The editor area is not functional, yet
shows the character being edited.

A new sink (SingleFrameSink) is introduced in this commit. Its goal is
to avoid the hassle of a separate thread for rendering such things as
single letters (perhaps we should adopt this sink in the editor widget
rather than the full-blown observer sink).

An important addition in this commit is the Rasterizer, which has spread
around everywhere in Siganim. The rasterizer's role is to draw pixels
according to some drawing pattern (for instance, Helen's style, Edigir's
style, etc). The user will be -is not yet- able to create those patterns
out of PNG files (and why not draw them directly in a later stage). This
will make Siganim a truly generic sign renderer, which will be able to
create signs in any commercial editor's style.

This commit leaves lots of work to be done regarding memory management,
and probably lots of room for improvement on conceptual design regarding
the font studio (say, is the Qt model for a font adapted to the use we
are making of it? Aren't we hijacking it?). I also don't like the "Text"
cell header inclusion all around user interface components, as well as
the Rasterizer being spread everywhere. It however looks like for the
rasterizer that we've got to either accept its inclusion in all UI
elements leading to rendered images to pass it from outside Qt, or we
have to use globals (perhaps SiganimDefaults can help).