First working version of font studio, RGBA colors
This commit brings in a font studio that "just works". The pixel scaling
factor is hardcoded to be 10, yet this is going to be changed by having
the rasterizer compute the pixel horizontal and vertical sizes.

SignColors that were RGB colors have become RGBA colors; the added alpha
channel is not yet of any use, but an extension of the Compose operation
can make it useful. Compose could indeed be extended to support add,
sub, mul, xor, and, or operations instead of the only current option
(or).