Fix more memory leaks - no more render-induced leak
This commit fixes several memory leaks noticed throughout the execution
of the program. The big fish (memory leak on render) has been cleared;
it was due to multiple allocations of the pixel area in the SignImage
that results from a composition (nested constructor calls triggered
several allocations, only the last of which was freed).

There remains some unclear areas on object deletion, notably observable
sinks that may happen to call for their sign to detach after it has been
deleted. Sign ownership issues may have to be worked out before this
gets solved.