Add JSON file support, implement JSON font support
This commit brings in a new file format for Siganim data, which is based
on JSON. Anything that can be dumped to JSON format should implement the
Serializable class.

Font objects are supported, and merely dumped with their existing
structure into this container format that suits it pretty much. Font
bitmap data is encoded as base64 then dumped as a string. There is no
compression though (that would be a nice addition).