Work on Rasterizer, logo, hardcode Unicode blocks, SiganimCore/UICore
This commit brings in SiganimCore and SiganiumUICore classes, which are
intended to contain the sign / fonts / rasterizers, along with some link
to be defined with the file they come from (local database, or user's
file, or both). The UI is handled by SiganimUICore, that's aimed at
containing the most common UI functions (those that can be accessed
through the menus, for instance).

Font Studio has support for (presumably) all Unicode printable
characters to date; yet, even though there is a way to enumerate all
Unicode blocks from ICU, there looks like to be no way to get ICU to
tell where a block starts and ends, making it impossible to build up a
list of a block's characters without browsing through all the preceding
ones. For this reason, the blocks and their boundaries are hardcoded in
a header file.

Rasterizers now have two metric features: their own height/width, and a
scale factor that's given by the user. The rasterizer's rendering
function is supposed to scale its output by the scale factor. As there
is a single kind of rasterizer, this property is honored as of now.

This commit also introduces an embryo of Siganim JSON sign file parser,
which files will contain our IR just as fonts do. Renderers will get the
same eventually.

A bit of memory cleanup has been added, we're almost done with memory
leaks. I however introduced a source of leaks I know about, which is in
the font studio: the model will generate an option structure for every
character that it will embed in a ModelIndex, and that option structure
never gets deleted. Keeping track of them should alleviate the issue,
now we have to do it properly.

Finally, in the perspective of writing a bit of documentation, I have
drawn a logo for Siganim, and derived an icon for that. They are
introduced in this commit, but not used anywhere in the program until I
find a way to embed resources that does not involve QMake (perhaps
dumping them as arrays will do it, though a bit of time is needed to
figure out how).