Add JSON sign input/output facilities
This commit brings in sign I/O under the form of JSON objects, just like
fonts. This is a first attempt and the separation of the parser and the
code generator (toJSON() / buildXXX()) is probably not the best option.
Notably, we are using string constants for property names and
enumerables, which have to be kept consistent between codegen and
parser.

A sample JSON sign is included, along with a sample JSON font file, to
be used as a starting point by new users.