Add a widget to create a new character into the Font Studio
This commit brings in a widget that is shown whenever a Unicode point
that does not belong to the currently selected font is picked. This
widget contains an "Add" button to create that character into the font.

It would probably be a good idea for the Font Studio to only keep a list
of those characters in the font and have a separate dialog to add new
ones; this dialog would then show Unicode blocks and unavailable
characters (thus getting rid of character options and of this commit).