Switch to CMake, Siganim now builds for Windows
This commit switches build systems, now relying on CMake instead of
Eclipse CDT's managed builder.

Using MXE, we can build Siganim with either static or dynamic linking; a
static executable is about 73MB (closer to 50 when stripped), whereas a
dynamic one is about 3MB along with a 90MB DLL package.

Two switches are available for this project:

  - STATIC_LINKING, that instructs the build system to use static
versions of ICU (which names begin with an "s") - defaults to OFF,
  - GUI_ENABLE, that builds Siganim with its Qt GUI - defaults to ON.
Disabling this option removes the need for Qt.